feat: refactor audits and dashboard components for improved structure and user experience; add ActiveScanBadge and AuditDetailCard components
