+++ src/app/(general)/audits/page.tsx
@@ -2,30 +2,27 @@ import { safeAwait } from '@/lib/async'
 import { BasicAlert } from '@/components/common/BasicAlert'
 import { getSession } from '@/lib/session'
 import { getAuditsForUser } from '@/actions/audits/getAuditList/logic'
-import { AuditsContainer } from '@/components/pages/audits/AuditsContainer'
+import { AuditsContainer } from '@/components/pages/audits/AuditsPageContainer'
+import { redirect } from 'next/navigation'
 
 export default async function AuditsPage() {
   const session = await getSession()
   const userId = session.user?.id
 
   if (!userId) {
-    return (
-      <BasicAlert
-        variant="destructive"
-        title="Not authenticated"
-        description="Please sign in to view your audits."
-      />
-    )
+    redirect('/signin')
   }
 
   const [err, res] = await safeAwait(getAuditsForUser(userId))
   if (err || !res?.success) {
     return (
-      <BasicAlert
-        variant="destructive"
-        title="Error loading audits"
-        description="There was a problem fetching your audits. Please try again."
-      />
+      <div className='p-10'> 
+          <BasicAlert
+            variant="destructive"
+            title="Error loading audits"
+            description="There was a problem fetching your audits. Please try again."
+          />
+      </div>
     )
   }
